Narrow Codex install config edits
