+++ src/install.ts
@@ -1,5 +1,6 @@
-import { mkdir, readFile, writeFile } from "node:fs/promises";
+import { mkdir, writeFile } from "node:fs/promises";
 import { dirname, join } from "node:path";
+import { configureCodexInstall } from "./adapters/codex-config";
 import { installCodexHomeProjection, installPiDevHomeProjection } from "./home-projection";
 import { buildSomaStartupContext, runSomaLifecycleAlgorithmUpdated } from "./lifecycle";
 import { bootstrapSomaHome } from "./soma-home";
@@ -119,181 +120,6 @@ async function installSomaForSubstrate(
       files: [...substrateHome.files, ...configFiles, ...lifecycleFiles],
     },
   };
-}
-
-async function configureCodexInstall(codexHome: string, somaHome: string): Promise<string> {
-  const path = join(codexHome, "config.toml");
-  const existing = await readFile(path, "utf8").catch(() => "");
-  let next = enableCodexHooksFeature(existing);
-  next = enableCodexWorkspaceWrite(next, somaHome);
-
-  await mkdir(dirname(path), { recursive: true });
-  await writeFile(path, next.trimEnd() + "\n", "utf8");
-
-  return path;
-}
-
-function enableCodexHooksFeature(config: string): string {
-  let next = config;
-
-  if (!/^\[features\]$/m.test(next)) {
-    next = `${next.trimEnd()}\n\n[features]\nhooks = true\n`;
-  } else if (/^codex_hooks\s*=/m.test(next)) {
-    next = next.replace(/^codex_hooks\s*=.*$/m, "hooks = true");
-  } else if (!/^hooks\s*=/m.test(next)) {
-    next = next.replace(/^\[features\]$/m, "[features]\nhooks = true");
-  } else {
-    next = next.replace(/^hooks\s*=.*$/m, "hooks = true");
-  }
-
-  return next;
-}
-
-function enableCodexWorkspaceWrite(config: string, somaHome: string): string {
-  let next = ensureTopLevelCodexSandboxMode(config);
-  next = upsertCodexWritableRoot(next, somaHome);
-
-  return next;
-}
-
-function ensureTopLevelCodexSandboxMode(config: string): string {
-  const topLevel = config.slice(0, findFirstTomlSectionStart(config));
-
-  if (findTomlKey(topLevel, "sandbox_mode") !== undefined) {
-    return config;
-  }
-
-  const trimmed = config.trimStart();
-  const leading = config.slice(0, config.length - trimmed.length);
-  return `${leading}sandbox_mode = "workspace-write"\n\n${trimmed}`;
-}
-
-function upsertCodexWritableRoot(config: string, somaHome: string): string {
-  const section = findTomlSection(config, "sandbox_workspace_write");
-
-  if (section === undefined) {
-    return `${config.trimEnd()}\n\n[sandbox_workspace_write]\nwritable_roots = [${quoteTomlString(somaHome)}]\n`;
-  }
-
-  const body = config.slice(section.bodyStart, section.bodyEnd);
-  const key = findTomlKey(body, "writable_roots");
-
-  if (key === undefined) {
-    const insertAt = section.headerEnd;
-    return `${config.slice(0, insertAt)}writable_roots = [${quoteTomlString(somaHome)}]\n${config.slice(insertAt)}`;
-  }
-
-  const roots = parseTomlStringArray(key.value);
-  if (!roots.includes(somaHome)) {
-    roots.push(somaHome);
-  }
-
-  const replacement = `writable_roots = [${roots.map(quoteTomlString).join(", ")}]`;
-  const start = section.bodyStart + key.start;
-  const end = section.bodyStart + key.end;
-  return `${config.slice(0, start)}${replacement}${config.slice(end)}`;
-}
-
-function findTomlSection(config: string, name: string): { bodyStart: number; bodyEnd: number; headerEnd: number } | undefined {
-  const headerPattern = new RegExp(`^\\[${escapeRegExp(name)}\\]\\s*$`, "m");
-  const header = headerPattern.exec(config);
-
-  if (header === null) {
-    return undefined;
-  }
-
-  const headerEnd = header.index + header[0].length + (config[header.index + header[0].length] === "\n" ? 1 : 0);
-  const rest = config.slice(headerEnd);
-  const nextHeader = /^\[[^\]]+\]\s*$/m.exec(rest);
-  const bodyEnd = nextHeader?.index === undefined ? config.length : headerEnd + nextHeader.index;
-
-  return {
-    bodyStart: headerEnd,
-    bodyEnd,
-    headerEnd,
-  };
-}
-
-function findFirstTomlSectionStart(config: string): number {
-  const section = /^\[[^\]]+\]\s*$/m.exec(config);
-  return section?.index ?? config.length;
-}
-
-function findTomlKey(config: string, key: string): { start: number; end: number; value: string } | undefined {
-  const keyPattern = new RegExp(`^\\s*${escapeRegExp(key)}\\s*=\\s*`, "m");
-  const match = keyPattern.exec(config);
-
-  if (match === null) {
-    return undefined;
-  }
-
-  const valueStart = match.index + match[0].length;
-  const end = config[valueStart] === "[" ? findTomlArrayEnd(config, valueStart) : findTomlLineEnd(config, valueStart);
-
-  return {
-    start: match.index,
-    end,
-    value: config.slice(valueStart, end).trim(),
-  };
-}
-
-function findTomlArrayEnd(config: string, start: number): number {
-  let inString = false;
-  let escaped = false;
-  let depth = 0;
-
-  for (let index = start; index < config.length; index += 1) {
-    const char = config[index];
-
-    if (inString) {
-      if (escaped) {
-        escaped = false;
-      } else if (char === "\\") {
-        escaped = true;
-      } else if (char === '"') {
-        inString = false;
-      }
-      continue;
-    }
-
-    if (char === '"') {
-      inString = true;
-    } else if (char === "[") {
-      depth += 1;
-    } else if (char === "]") {
-      depth -= 1;
-      if (depth === 0) {
-        return index + 1;
-      }
-    }
-  }
-
-  return findTomlLineEnd(config, start);
-}
-
-function findTomlLineEnd(config: string, start: number): number {
-  const newline = config.indexOf("\n", start);
-  return newline === -1 ? config.length : newline;
-}
-
-function parseTomlStringArray(value: string): string[] {
-  const roots: string[] = [];
-  const stringPattern = /"((?:\\.|[^"\\])*)"/g;
-  let match: RegExpExecArray | null;
-
-  while ((match = stringPattern.exec(value)) !== null) {
-    roots.push(match[1].replace(/\\"/g, '"').replace(/\\\\/g, "\\"));
-  }
-
-  return roots;
-}
-
-function quoteTomlString(value: string): string {
-  return `"${value.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`;
-}
-
-function escapeRegExp(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
 }
 
 async function writeProjectionFile(root: string, relativePath: string, content: string): Promise<string> {
